Fixing new placement code from Karlsruhe (Bug #2649)
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@10885 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/tool/placement/genetic1/Population.java
@@ -53,9 +53,9 @@ public class Population {
 		double best_fitness = Double.MAX_VALUE;
 		
 		for (Chromosome c : chromosomes) {
-			assert (c.fitness > 0);
-			if (c.fitness < best_fitness)
-				best_fitness = c.fitness;
+			assert (c.fitness.doubleValue() > 0);
+			if (c.fitness.doubleValue() < best_fitness)
+				best_fitness = c.fitness.doubleValue();
 		}
 		
 		return best_fitness;
@@ -66,7 +66,7 @@ public class Population {
 
 		// place altered cells before calculating metric
 		for (Chromosome c : chromosomes) {
-			assert (c.altered || c.fitness > 0);
+			assert (c.altered || c.fitness.doubleValue() > 0);
 			placement.placeChromosome(c);
 		}
